unit_two/requirements.txt
click==8.1.3
colorama==0.4.5
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.3
pandas==1.5.0
psycopg2-binary==2.9.4
python-dateutil==2.8.2
pytz==2022.4
six==1.16.0
Werkzeug==2.2.2
